changing package namespace to edu.washington.cirg.*, updating pom to publish classes jar to local maven repo, changing docker-build.bat to tag image with uwcirg image name
