+++ src/main/java/edu/washington/cirg/fhir/jpa/starter/FhirServerConfigR5.java
@@ -1,7 +1,7 @@
-package ca.uhn.fhir.jpa.starter;
+package edu.washington.cirg.fhir.jpa.starter;
 
 import ca.uhn.fhir.context.ConfigurationException;
-import ca.uhn.fhir.jpa.config.BaseJavaConfigDstu3;
+import ca.uhn.fhir.jpa.config.BaseJavaConfigR5;
 import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -13,7 +13,7 @@ import javax.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
 @Configuration
-public class FhirServerConfigDstu3 extends BaseJavaConfigDstu3 {
+public class FhirServerConfigR5 extends BaseJavaConfigR5 {
 
     @Autowired
     private DataSource myDataSource;
